Conflicts:
	corehq/apps/registration/views.py

+++ custom/ilsgateway/tasks.py
@@ -9,6 +9,7 @@ from django.db import transaction
 from psycopg2._psycopg import DatabaseError
 
 from corehq.apps.locations.models import SQLLocation
+from custom.ilsgateway.slab.reminders.stockout import StockoutReminder
 from custom.ilsgateway.tanzania.reminders import REMINDER_MONTHLY_SOH_SUMMARY, REMINDER_MONTHLY_DELIVERY_SUMMARY, \
     REMINDER_MONTHLY_RANDR_SUMMARY
 from custom.ilsgateway.tanzania.reminders.delivery import DeliveryReminder
@@ -122,7 +123,6 @@ def report_run_periodic_task():
 @task(queue='logistics_background_queue', ignore_result=True)
 def report_run(domain, locations=None, strict=True):
     last_successful_run = ReportRun.last_success(domain)
-    recalculation_on_location_change(domain, last_successful_run)
 
     last_run = ReportRun.last_run(domain)
     start_date = (datetime.min if not last_successful_run else last_successful_run.end)
@@ -163,6 +163,8 @@ def report_run(domain, locations=None, strict=True):
         run.complete = True
         run.save()
         logging.info("ILSGateway report runner end time: %s" % datetime.utcnow())
+        if not has_error:
+            recalculation_on_location_change.delay(domain, last_successful_run)
 
 facility_delivery_partial = partial(send_for_day, cutoff=15, reminder_class=DeliveryReminder)
 district_delivery_partial = partial(send_for_day, cutoff=13, reminder_class=DeliveryReminder,
@@ -360,13 +362,26 @@ def soh_thank_you_task():
         SOHThankYouReminder(domain=domain, date=last_month).send()
 
 
+@periodic_task(run_every=crontab(day_of_month="6-10", hour=8, minute=0),
+               queue="logistics_reminder_queue")
+def stockout_reminder_task():
+    """
+        6th business day of month
+    """
+    now = datetime.utcnow()
+    last_business_day = get_business_day_of_month(month=now.month, year=now.year, count=6)
+    if now.day != last_business_day.day:
+        return
+
+    send_for_all_domains(last_business_day, StockoutReminder)
+
+
 def recalculate_on_group_change(location, last_run):
     OrganizationSummary.objects.filter(location_id=location.get_id).delete()
     process_facility_warehouse_data(location, default_start_date(), last_run.end)
-
-    for parent in location.sql_location.get_ancestors(ascending=True):
-        process_non_facility_warehouse_data(parent.couch_location,
-                                            default_start_date(), last_run.end, strict=False)
+    return {
+        ancestor.location_type.name: {ancestor} for ancestor in location.sql_location.get_ancestors(ascending=True)
+    }
 
 
 def recalculate_on_parent_change(location, previous_parent_id, last_run):
@@ -387,13 +402,10 @@ def recalculate_on_parent_change(location, previous_parent_id, last_run):
     for sql_location in locations_to_recalculate:
         type_location_map[sql_location.location_type.name].add(sql_location)
 
-    for location_type in ["DISTRICT", "REGION", "MSDZONE"]:
-        for sql_location in type_location_map[location_type]:
-            process_non_facility_warehouse_data(
-                sql_location.couch_location, default_start_date(), last_run.end, strict=False
-            )
+    return type_location_map
 
 
+@task(queue='logistics_background_queue', ignore_result=True)
 def recalculation_on_location_change(domain, last_run):
     if not last_run:
         PendingReportingDataRecalculation.objects.filter(domain=domain).delete()
@@ -406,19 +418,42 @@ def recalculation_on_location_change(domain, last_run):
         key = (pending_recalculation.sql_location, pending_recalculation.type)
         recalcs_dict[key].append(pending_recalculation.data)
 
+    non_facilities_to_recalculate = defaultdict(set)
+    recalculated = set()
     for (sql_location, recalculation_type), data_list in recalcs_dict.iteritems():
         # If there are more changes, consider earliest and latest change.
         # Thanks to this we avoid recalculations when in fact group/parent wasn't changed.
         # E.g Group is changed from A -> B and later from B -> A.
         # In this situation there is no need to recalculate data.
+
+        if not OrganizationSummary.objects.filter(location_id=sql_location.location_id).exists():
+            # There are no data for that location so there is no need to recalculate
+            PendingReportingDataRecalculation.objects.filter(
+                sql_location=sql_location, type=recalculation_type, domain=domain
+            ).delete()
+            continue
+
         if recalculation_type == 'group_change'\
                 and data_list[0]['previous_group'] != data_list[-1]['current_group']:
-            recalculate_on_group_change(sql_location.couch_location, last_run)
+            to_recalculate = recalculate_on_group_change(sql_location.couch_location, last_run)
         elif recalculation_type == 'parent_change' \
                 and data_list[0]['previous_parent'] != data_list[-1]['current_parent']:
-            recalculate_on_parent_change(
+            to_recalculate = recalculate_on_parent_change(
                 sql_location.couch_location, data_list[0]['previous_parent'], last_run
             )
-        PendingReportingDataRecalculation.objects.filter(
-            sql_location=sql_location, type=recalculation_type, domain=domain
-        ).delete()
+        else:
+            to_recalculate = {}
+        recalculated.add(sql_location)
+
+        for location_type, sql_locations_to_recalculate in to_recalculate.iteritems():
+            for sql_location_to_recalculate in sql_locations_to_recalculate:
+                non_facilities_to_recalculate[location_type].add(sql_location_to_recalculate)
+
+    for location_type in ["DISTRICT", "REGION", "MSDZONE", "MOHSW"]:
+        for sql_location in non_facilities_to_recalculate.get(location_type, []):
+            process_non_facility_warehouse_data(
+                sql_location.couch_location, default_start_date(), last_run.end, strict=False
+            )
+    PendingReportingDataRecalculation.objects.filter(
+        sql_location__in=recalculated, domain=domain
+    ).delete()
